+++ applications/fishing-map/src/features/analysis/Analysis.module.css
@@ -42,17 +42,6 @@
   margin-bottom: auto;
 }
 
-.commonTitle {
-  width: 100%;
-  padding: var(--space-M);
-  border-bottom: var(--border);
-}
-
-.layerPanels {
-  display: flex;
-  flex-wrap: wrap;
-}
-
 .saveBtn {
   margin-top: auto;
   align-self: flex-end;
@@ -63,11 +52,6 @@
   font: var(--font-S);
   color: var(--color-secondary-blue);
   padding: var(--space-M);
-}
-
-.graph {
-  padding: var(--space-M);
-  min-height: 28rem;
 }
 
 .graph .spinner {
