+++ applications/fishing-map/src/features/analysis/AnalysisItem.module.css
@@ -1,0 +1,21 @@
+.container {
+  margin-bottom: 5rem;
+  border-top: var(--border);
+}
+
+.commonTitle {
+  width: 100%;
+  padding: var(--space-M);
+  border-bottom: var(--border);
+}
+
+.layerPanels {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.placeholder {
+  font: var(--font-S);
+  color: var(--color-secondary-blue);
+  padding: var(--space-M);
+}
